Moves files to correct folders
